edi_storage: add base tests for component
