+++ edi_storage_oca/tests/__init__.py
@@ -1,1 +1,2 @@
 from . import test_edi_backend_storage
+from . import test_components_base
